Support the new color struct
The C.VTermColor struct is now a union, representing either a RGB color or a color
index. This updated code exposes this new API: the color type can be found
with IsRGB() and IsIndex(), and the value can be get with GetRGB() and
GetIndex().

This update also binds vterm_state_convert_color_to_rgb.

This update breaks the pervious API for getting color -- cell.Fg() and
cell.Bg() now returns VTermColor, which does not implement color.Color.
This is because it is not convenient to determine the RGB value of an
indexed color without a reference to the state. To do that, call
state.ConvertVTermColorToRGB.